Move utility method.
