+++ src/main/java/com/powsybl/openloadflow/network/impl/HvdcConverterStations.java
@@ -8,6 +8,7 @@ package com.powsybl.openloadflow.network.impl;
 
 import com.powsybl.iidm.network.*;
 import com.powsybl.iidm.network.util.HvdcUtils;
+import com.powsybl.openloadflow.network.LfNetworkParameters;
 
 import java.util.Optional;
 
@@ -39,5 +40,29 @@ public final class HvdcConverterStations {
     public static boolean isVsc(Connectable<?> connectable) {
         return connectable.getType() == IdentifiableType.HVDC_CONVERTER_STATION
                 && ((HvdcConverterStation<?>) connectable).getHvdcType() == HvdcConverterStation.HvdcType.VSC;
+    }
+
+    public static boolean isHvdcDanglingInIidm(HvdcConverterStation<?> station, LfNetworkParameters parameters) {
+
+        if (isIsolated(station.getTerminal(), parameters)) {
+            return true;
+        } else {
+            return station.getOtherConverterStation().map(otherConverterStation -> {
+                Terminal otherTerminal = otherConverterStation.getTerminal();
+                return isIsolated(otherTerminal, parameters);
+            }).orElse(true); // it means there is no HVDC line connected to station
+        }
+    }
+
+    private static boolean isIsolated(Terminal terminal, LfNetworkParameters parameters) {
+        Bus bus = parameters.isBreakers() ? terminal.getBusBreakerView().getBus() : terminal.getBusView().getBus();
+        if (bus == null) {
+            return true;
+        }
+
+        // The criteria should as close as possible to Networks.isIsolatedBusForHvdc - only connected to the station
+        return bus.getConnectedTerminalStream()
+                .map(Terminal::getConnectable)
+                .noneMatch(c -> !(c instanceof HvdcConverterStation<?> || c instanceof BusbarSection));
     }
 }
