+++ src/main/java/com/powsybl/openloadflow/network/impl/LfVscConverterStationImpl.java
@@ -16,7 +16,6 @@ import com.powsybl.openloadflow.network.LfNetwork;
 import com.powsybl.openloadflow.network.LfNetworkParameters;
 import com.powsybl.openloadflow.network.LfNetworkStateUpdateParameters;
 import com.powsybl.openloadflow.network.LfVscConverterStation;
-import com.powsybl.openloadflow.util.LfHvdcUtils;
 import com.powsybl.openloadflow.util.PerUnit;
 
 import java.util.Objects;
@@ -37,7 +36,7 @@ public class LfVscConverterStationImpl extends AbstractLfGenerator implements Lf
 
     public LfVscConverterStationImpl(VscConverterStation station, LfNetwork network, LfNetworkParameters parameters, LfNetworkLoadingReport report) {
         super(network, HvdcUtils.getConverterStationTargetP(station) / PerUnit.SB);
-        this.hvdcDandlingInIidm = LfHvdcUtils.isHvdcDanglingInIidm(station, parameters);
+        this.hvdcDandlingInIidm = HvdcConverterStations.isHvdcDanglingInIidm(station, parameters);
         this.stationRef = Ref.create(station, parameters.isCacheEnabled());
         this.lossFactor = station.getLossFactor();
